requirements.txt
Flask==2.2.2
Flask-Caching==2.0.1
Flask-DotEnv==0.1.2
python-dotenv==0.21.0
redis==4.3.4
requests==2.28.1

